Remove AutoLock tab, move options to general tab
