+++ src/main/java/org/cryptomator/ui/vaultoptions/GeneralVaultOptionsController.java
@@ -3,10 +3,11 @@ package org.cryptomator.ui.vaultoptions;
 import org.cryptomator.common.settings.WhenUnlocked;
 import org.cryptomator.common.vaults.Vault;
 import org.cryptomator.ui.common.FxController;
-import org.cryptomator.ui.health.HealthCheckComponent;
+import org.cryptomator.ui.controls.NumericTextField;
 
 import javax.inject.Inject;
 import javafx.beans.Observable;
+import javafx.beans.binding.Bindings;
 import javafx.fxml.FXML;
 import javafx.scene.control.CheckBox;
 import javafx.scene.control.ChoiceBox;
@@ -28,9 +29,11 @@ public class GeneralVaultOptionsController implements FxController {
 	public TextField vaultName;
 	public CheckBox unlockOnStartupCheckbox;
 	public ChoiceBox<WhenUnlocked> actionAfterUnlockChoiceBox;
+	public CheckBox lockAfterTimeCheckbox;
+	public NumericTextField lockTimeInMinutesTextField;
 
 	@Inject
-	GeneralVaultOptionsController(@VaultOptionsWindow Stage window, @VaultOptionsWindow Vault vault,  ResourceBundle resourceBundle) {
+	GeneralVaultOptionsController(@VaultOptionsWindow Stage window, @VaultOptionsWindow Vault vault, ResourceBundle resourceBundle) {
 		this.window = window;
 		this.vault = vault;
 		this.resourceBundle = resourceBundle;
@@ -45,6 +48,8 @@ public class GeneralVaultOptionsController implements FxController {
 		actionAfterUnlockChoiceBox.getItems().addAll(WhenUnlocked.values());
 		actionAfterUnlockChoiceBox.valueProperty().bindBidirectional(vault.getVaultSettings().actionAfterUnlock());
 		actionAfterUnlockChoiceBox.setConverter(new WhenUnlockedConverter(resourceBundle));
+		lockAfterTimeCheckbox.selectedProperty().bindBidirectional(vault.getVaultSettings().autoLockWhenIdle());
+		Bindings.bindBidirectional(lockTimeInMinutesTextField.textProperty(), vault.getVaultSettings().autoLockIdleSeconds(), new IdleTimeSecondsConverter());
 	}
 
 	private void trimVaultNameOnFocusLoss(Observable observable, Boolean wasFocussed, Boolean isFocussed) {
@@ -81,4 +86,22 @@ public class GeneralVaultOptionsController implements FxController {
 		}
 	}
 
+	private static class IdleTimeSecondsConverter extends StringConverter<Number> {
+
+		@Override
+		public String toString(Number seconds) {
+			int minutes = seconds.intValue() / 60; // int-truncate
+			return Integer.toString(minutes);
+		}
+
+		@Override
+		public Number fromString(String string) {
+			try {
+				int minutes = Integer.valueOf(string);
+				return minutes * 60;
+			} catch (NumberFormatException e) {
+				return 0;
+			}
+		}
+	}
 }
