+++ src/main/resources/fxml/vault_options.fxml
@@ -36,14 +36,6 @@
 				<fx:include source="/fxml/vault_options_masterkey.fxml"/>
 			</content>
 		</Tab>
-		<Tab fx:id="autoLockTab" id="AUTOLOCK" text="%vaultOptions.autoLock">
-			<graphic>
-				<FontAwesome5IconView glyph="STOPWATCH"/>
-			</graphic>
-			<content>
-				<fx:include source="/fxml/vault_options_autolock.fxml"/>
-			</content>
-		</Tab>
 		<Tab fx:id="healthTab" id="HEALTH" text="%vaultOptions.health">
 			<graphic>
 				<FontAwesome5IconView glyph="STETHOSCOPE"/>
