--- src/main/resources/fxml/vault_options_autolock.fxml
@@ -1,26 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-
-<?import org.cryptomator.ui.controls.FormattedLabel?>
-<?import org.cryptomator.ui.controls.NumericTextField?>
-<?import javafx.geometry.Insets?>
-<?import javafx.scene.control.CheckBox?>
-<?import javafx.scene.layout.VBox?>
-<?import javafx.scene.text.Text?>
-<?import javafx.scene.text.TextFlow?>
-<VBox xmlns:fx="http://javafx.com/fxml"
-	  xmlns="http://javafx.com/javafx"
-	  fx:controller="org.cryptomator.ui.vaultoptions.AutoLockVaultOptionsController"
-	  spacing="6">
-	<padding>
-		<Insets topRightBottomLeft="12"/>
-	</padding>
-	<children>
-		<TextFlow styleClass="text-flow" prefWidth="-Infinity">
-			<CheckBox text="%vaultOptions.autoLock.lockAfterTimePart1" fx:id="lockAfterTimeCheckbox"/>
-			<Text text=" "/>
-			<NumericTextField fx:id="lockTimeInMinutesTextField" prefWidth="50"/>
-			<Text text=" "/>
-			<FormattedLabel format="%vaultOptions.autoLock.lockAfterTimePart2"/>
-		</TextFlow>
-	</children>
-</VBox>
