+++ src/main/resources/fxml/vault_options_general.fxml
@@ -8,6 +8,10 @@
 <?import javafx.scene.layout.HBox?>
 <?import javafx.scene.layout.VBox?>
 <?import javafx.scene.control.Button?>
+<?import javafx.scene.text.TextFlow?>
+<?import javafx.scene.text.Text?>
+<?import org.cryptomator.ui.controls.NumericTextField?>
+<?import org.cryptomator.ui.controls.FormattedLabel?>
 <VBox xmlns:fx="http://javafx.com/fxml"
 	  xmlns="http://javafx.com/javafx"
 	  fx:controller="org.cryptomator.ui.vaultoptions.GeneralVaultOptionsController"
@@ -20,6 +24,14 @@
 			<Label text="%vaultOptions.general.vaultName"/>
 			<TextField fx:id="vaultName"/>
 		</HBox>
+
+		<TextFlow styleClass="text-flow" prefWidth="-Infinity">
+			<CheckBox text="%vaultOptions.general.autoLock.lockAfterTimePart1" fx:id="lockAfterTimeCheckbox"/>
+			<Text text=" "/>
+			<NumericTextField fx:id="lockTimeInMinutesTextField" prefWidth="50"/>
+			<Text text=" "/>
+			<FormattedLabel format="%vaultOptions.general.autoLock.lockAfterTimePart2"/>
+		</TextFlow>
 
 		<CheckBox text="%vaultOptions.general.unlockAfterStartup" fx:id="unlockOnStartupCheckbox"/>
