+++ src/main/resources/i18n/strings.properties
@@ -300,6 +300,8 @@ wrongFileAlert.link=For further assistance, visit
 ## General
 vaultOptions.general=General
 vaultOptions.general.vaultName=Vault Name
+vaultOptions.general.autoLock.lockAfterTimePart1=Lock when idle for
+vaultOptions.general.autoLock.lockAfterTimePart2=minutes
 vaultOptions.general.unlockAfterStartup=Unlock vault when starting Cryptomator
 vaultOptions.general.actionAfterUnlock=After successful unlock
 vaultOptions.general.actionAfterUnlock.ignore=Do nothing
@@ -323,10 +325,6 @@ vaultOptions.masterkey.forgetSavedPasswordBtn=Forget Saved Password
 vaultOptions.masterkey.recoveryKeyExpanation=A recovery key is your only means to restore access to a vault if you lose your password.
 vaultOptions.masterkey.showRecoveryKeyBtn=Display Recovery Key
 vaultOptions.masterkey.recoverPasswordBtn=Recover Password
-## Auto Lock
-vaultOptions.autoLock=Auto-Lock
-vaultOptions.autoLock.lockAfterTimePart1=Lock when idle for
-vaultOptions.autoLock.lockAfterTimePart2=minutes
 ## Health
 vaultOptions.health=Vault Health
 vaultOptions.health.startBtn=Start Health Check
